% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootwrq_cov.R
\name{bootwrq_cov}
\alias{bootwrq_cov}
\title{Title}
\usage{
bootwrq_cov(
  B,
  form,
  tau,
  data,
  Y,
  X1 = NULL,
  X2 = NULL,
  subject,
  death,
  time,
  interval.death = NULL,
  impute = NULL,
  weight = NULL,
  wcompute = 2,
  seed = NULL,
  intermittent,
  file = NULL,
  nproc = 1,
  MPI = FALSE
)
}
\description{
Title
}
