% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewrq.R
\name{.onewrq}
\alias{.onewrq}
\title{Title}
\usage{
.onewrq(
  form,
  tau,
  data,
  Y,
  X1,
  X2,
  subject,
  death,
  time,
  interval.death,
  impute,
  weight,
  wcompute,
  seed,
  intermittent
)
}
\description{
Title
}
