% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_reg_quant.R
\name{MSE_reg}
\alias{MSE_reg}
\title{Title}
\usage{
MSE_reg(
  traindata,
  newdata,
  tau,
  level = 0.05,
  N_obs,
  N_min,
  Y,
  spline,
  knots,
  time = NULL,
  X1,
  X2,
  subject,
  death,
  wcompute,
  intermittent,
  weight,
  B = 300,
  seed = floor(runif(300) * 1e+06),
  form_time
)
}
\description{
Title
}
