% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_WQ_solo.R
\name{predict_WQ_solo}
\alias{predict_WQ_solo}
\title{Title}
\usage{
predict_WQ_solo(
  newdata,
  new_subject,
  B = 100,
  form,
  tau,
  data,
  Y,
  X1,
  subject,
  death = "death",
  time = "time",
  wcompute = 0,
  intermittent = T,
  weight_WQ,
  seed
)
}
\description{
Title
}
